AI_backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
langgraph>=0.0.15
openai>=1.0.0
gTTS>=2.3.1
python-multipart>=0.0.6
logging>=0.5.1.2
requests>=2.31.0
pymongo>=4.6.0
